--- public/img/media/galleries/190911_diada/index.html
@@ -1,186 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
-    <meta http-equiv="Content-type" content="text/html;charset=UTF-8" />
-    <link rel="shortcut icon" href="style/favicon.ico" />
-    <link rel="stylesheet" href="style/style.css" type="text/css" />
-    <link rel="stylesheet" href="pswp/photoswipe.css">
-    <link rel="stylesheet" href="pswp/default-skin/default-skin.css">
-    <script src="pswp/photoswipe.min.js"></script>
-    <script src="pswp/photoswipe-ui-default.min.js"></script>
-    <title>darktable gallery</title>
-  </head>
-  <body>
-    <div class="title">darktable gallery</div>
-    <div class="page">
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_055-thumb.webp" alt="img1" class="img" onclick="openSwipe(0)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_056-thumb.webp" alt="img2" class="img" onclick="openSwipe(1)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_059-thumb.webp" alt="img3" class="img" onclick="openSwipe(2)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_065-thumb.webp" alt="img4" class="img" onclick="openSwipe(3)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_066-thumb.webp" alt="img5" class="img" onclick="openSwipe(4)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_068-thumb.webp" alt="img6" class="img" onclick="openSwipe(5)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_069-thumb.webp" alt="img7" class="img" onclick="openSwipe(6)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_073-thumb.webp" alt="img8" class="img" onclick="openSwipe(7)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_074-thumb.webp" alt="img9" class="img" onclick="openSwipe(8)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-
-      <div><div class="dia">
-      <img src="Diada-2019_Berlin_photo-by-Carlos-Collado_088-thumb.webp" alt="img10" class="img" onclick="openSwipe(9)"/></div>
-      <h1>&nbsp;</h1>
-      &nbsp;</div>
-        <p style="clear:both;"></p>
-    </div>
-    <div class="footer">
-      <script language="JavaScript" type="text/javascript">
-      document.write("download all: <em>curl -O#  " + document.documentURI.replace( /\\/g, '/' ).replace( /\/[^\/]*$/, '' ) + "/img_[0000-0010].jpg</em>")
-      </script><br />
-      created with darktable 2.4.2
-    </div>
-    <div class="pswp" tabindex="-1" role="dialog" aria-hidden="true">
-        <div class="pswp__bg"></div>
-        <div class="pswp__scroll-wrap">
-            <div class="pswp__container">
-                <div class="pswp__item"></div>
-                <div class="pswp__item"></div>
-                <div class="pswp__item"></div>
-            </div>
-            <div class="pswp__ui pswp__ui--hidden">
-                <div class="pswp__top-bar">
-                    <div class="pswp__counter"></div>
-                    <button class="pswp__button pswp__button--close" title="Close (Esc)"></button>
-                    <button class="pswp__button pswp__button--share" title="Share"></button>
-                    <button class="pswp__button pswp__button--fs" title="Toggle fullscreen"></button>
-                    <button class="pswp__button pswp__button--zoom" title="Zoom in/out"></button>
-                    <div class="pswp__preloader">
-                        <div class="pswp__preloader__icn">
-                          <div class="pswp__preloader__cut">
-                            <div class="pswp__preloader__donut"></div>
-                          </div>
-                        </div>
-                   </div>
-                </div>
-                <div class="pswp__share-modal pswp__share-modal--hidden pswp__single-tap">
-                    <div class="pswp__share-tooltip"></div>
-                </div>
-                <button class="pswp__button pswp__button--arrow--left" title="Previous (arrow left)">
-                </button>
-                <button class="pswp__button pswp__button--arrow--right" title="Next (arrow right)">
-                </button>
-                <div class="pswp__caption">
-                    <div class="pswp__caption__center"></div>
-                </div>
-            </div>
-        </div>
-    </div>
-  </body>
-<script>
-var pswpElement = document.querySelectorAll('.pswp')[0];
-var items = [
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_055.webp",
-w: 1200,
-h: 800,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_055-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_056.webp",
-w: 1200,
-h: 800,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_056-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_059.webp",
-w: 800,
-h: 1200,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_059-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_065.webp",
-w: 1200,
-h: 800,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_065-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_066.webp",
-w: 1200,
-h: 800,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_066-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_068.webp",
-w: 800,
-h: 1200,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_068-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_069.webp",
-w: 1200,
-h: 800,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_069-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_073.webp",
-w: 800,
-h: 1200,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_073-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_074.webp",
-w: 1200,
-h: 800,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_074-thumb.webp",
-},
-{
-src: "Diada-2019_Berlin_photo-by-Carlos-Collado_088.webp",
-w: 1200,
-h: 800,
-msrc: "Diada-2019_Berlin_photo-by-Carlos-Collado_088-thumb.webp",
-},
-];
-function openSwipe(img)
-{
-    // define options (if needed)
-    var options = {
-          // optionName: 'option value'
-          index: img // start at first slide
-    };
-    var gallery = new PhotoSwipe( pswpElement, PhotoSwipeUI_Default, items, options);
-    gallery.init();
-}
-</script>
-</html>
