+++ src/components/gallery_list.js
@@ -1,0 +1,59 @@
+const images = [
+  [
+    {
+      thumb: '/img/media/galleries/190911_diada/055-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/055.jpg',
+      caption: 'Actuació al FAZ Atrium',
+      date: '11 Setembre 2019',
+    },
+    {
+      thumb: '/img/media/galleries/190911_diada/059-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/059.jpg',
+      caption: 'Actuació al FAZ Atrium',
+      date: '11 Setembre 2019',
+    },
+  ],
+  [
+    {
+      thumb: '/img/media/galleries/190911_diada/074-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/074.jpg',
+      caption: 'Second Gallery',
+      date: '--',
+    },
+    {
+      thumb: '/img/media/galleries/190911_diada/088-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/088.jpg',
+      caption: 'Second Gallery',
+      date: '--',
+    },
+  ],
+  [
+    {
+      thumb: '/img/media/galleries/190911_diada/065-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/065.jpg',
+      caption: 'Third Gallery',
+      date: '--',
+    },
+    {
+      thumb: '/img/media/galleries/190911_diada/066-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/066.jpg',
+      caption: 'Third Gallery',
+      date: '--',
+    },
+    {
+      thumb: '/img/media/galleries/190911_diada/066-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/066.jpg',
+      caption: 'Third Gallery',
+      date: '--',
+    },
+    {
+      thumb: '/img/media/galleries/190911_diada/066-thumb.jpg',
+      src: '/img/media/galleries/190911_diada/066.jpg',
+      caption: 'Third Gallery',
+      date: '--',
+    },
+
+  ],
+];
+
+export default images;
