--- src/components/lightbox_list.js
@@ -1,59 +1,0 @@
-const images = [
-  [
-    {
-      thumb: '/img/media/galleries/190911_diada/055-thumb.webp',
-      src: '/img/media/galleries/190911_diada/055.webp',
-      caption: 'Actuació al FAZ Atrium',
-      date: '11 Setembre 2019',
-    },
-    {
-      thumb: '/img/media/galleries/190911_diada/059-thumb.webp',
-      src: '/img/media/galleries/190911_diada/059.webp',
-      caption: 'Actuació al FAZ Atrium',
-      date: '11 Setembre 2019',
-    },
-  ],
-  [
-    {
-      thumb: '/img/media/galleries/190911_diada/074-thumb.webp',
-      src: '/img/media/galleries/190911_diada/074.webp',
-      caption: 'Second Gallery',
-      date: '--',
-    },
-    {
-      thumb: '/img/media/galleries/190911_diada/088-thumb.webp',
-      src: '/img/media/galleries/190911_diada/088.webp',
-      caption: 'Second Gallery',
-      date: '--',
-    },
-  ],
-  [
-    {
-      thumb: '/img/media/galleries/190911_diada/065-thumb.webp',
-      src: '/img/media/galleries/190911_diada/065.webp',
-      caption: 'Third Gallery',
-      date: '--',
-    },
-    {
-      thumb: '/img/media/galleries/190911_diada/066-thumb.webp',
-      src: '/img/media/galleries/190911_diada/066.webp',
-      caption: 'Third Gallery',
-      date: '--',
-    },
-    {
-      thumb: '/img/media/galleries/190911_diada/066-thumb.webp',
-      src: '/img/media/galleries/190911_diada/066.webp',
-      caption: 'Third Gallery',
-      date: '--',
-    },
-    {
-      thumb: '/img/media/galleries/190911_diada/066-thumb.webp',
-      src: '/img/media/galleries/190911_diada/066.webp',
-      caption: 'Third Gallery',
-      date: '--',
-    },
-
-  ],
-];
-
-export default images;
